requirements.txt
flask==3.0.0
langchain==0.3.0
langchain-openai==0.2.11
langchain-community==0.3.0
openai==1.57.4
python-dotenv==1.0.0
pydantic==2.10.5
requests==2.31.0
nest-asyncio==1.6.0
